mtd: ubi: add a ubi forced detach ioctl
When running ubi rootfs upgrade on nand based OpenWrt device after
pivot_root init process is still "hooked" on the old file system.
Thus, the old file system can not be umounted. If the filesystem is
mounted it can not be upgraded with using for example ubiupdatevol
or removed with ubirmvol. Forcing umount would allow to run the
before mentioned commands.

Signed-off-by: Sean Khan <datapronix@protonmail.com>